# backend/app/api/contracts/step2-unified-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TMDB 海报墙",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/collection/add": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "添加收藏",
        "description": "添加收藏\n\n- **tmdb_id**: TMDB ID\n- **media_type**: 媒体类型 (movie 或 tv)\n- **title**: 标题\n- **share_url**: 夸克分享链接",
        "operationId": "add_collection_api_collection_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionAddData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/list": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "获取收藏列表",
        "description": "获取收藏列表\n\n支持分页、排序和过滤",
        "operationId": "list_collections_api_collection_list_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序字段",
              "default": "saved_at",
              "title": "Sort By"
            },
            "description": "排序字段"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序方向: asc 或 desc",
              "default": "desc",
              "title": "Order"
            },
            "description": "排序方向: asc 或 desc"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类过滤",
              "title": "Category"
            },
            "description": "分类过滤"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "状态过滤",
              "title": "Status"
            },
            "description": "状态过滤"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionListData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/check/{tmdb_id}": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "检查是否已收藏",
        "description": "检查指定 TMDB ID 是否已收藏",
        "operationId": "check_collection_api_collection_check__tmdb_id__get",
        "parameters": [
          {
            "name": "tmdb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "TMDB ID",
              "title": "Tmdb Id"
            },
            "description": "TMDB ID"
          },
          {
            "name": "media_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "媒体类型",
              "default": "movie",
              "title": "Media Type"
            },
            "description": "媒体类型"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionCheckData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/{collection_id}": {
      "delete": {
        "tags": [
          "collection"
        ],
        "summary": "删除收藏",
        "description": "删除指定的收藏",
        "operationId": "delete_collection_api_collection__collection_id__delete",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "收藏 ID",
              "title": "Collection Id"
            },
            "description": "收藏 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionDeleteData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/check-link": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "检查链接是否已收藏",
        "description": "检查指定分享链接是否已收藏\n\n返回收藏状态和收藏 ID、状态",
        "operationId": "check_link_collection_api_collection_check_link_get",
        "parameters": [
          {
            "name": "link",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "分享链接",
              "title": "Link"
            },
            "description": "分享链接"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionCheckLinkData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/check-links": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "批量检查链接收藏状态",
        "description": "批量检查多个分享链接的收藏状态\n\n返回每个链接的收藏状态、收藏 ID 和转存状态",
        "operationId": "check_links_collection_api_collection_check_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCheckLinksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionCheckLinksData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/verify": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "验证收藏网盘状态（SSE）",
        "description": "验证收藏对应网盘目录是否存在，并同步修正收藏状态。",
        "operationId": "verify_collections_api_collection_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CollectionVerifyRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream. Each event payload is SseEnvelope JSON string prefixed with data: ",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/verify/{collection_id}": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "验证单个收藏网盘状态",
        "operationId": "verify_single_collection_api_collection_verify__collection_id__post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "收藏 ID",
              "title": "Collection Id"
            },
            "description": "收藏 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CollectionVerifySingleData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse_CollectionAddData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionAddData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionAddData]"
      },
      "ApiResponse_CollectionCheckData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionCheckData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionCheckData]"
      },
      "ApiResponse_CollectionCheckLinkData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionCheckLinkData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionCheckLinkData]"
      },
      "ApiResponse_CollectionCheckLinksData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionCheckLinksData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionCheckLinksData]"
      },
      "ApiResponse_CollectionDeleteData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionDeleteData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionDeleteData]"
      },
      "ApiResponse_CollectionListData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionListData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionListData]"
      },
      "ApiResponse_CollectionVerifySingleData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Business code. 0 means success."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable message."
          },
          "data": {
            "$ref": "#/components/schemas/CollectionVerifySingleData"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "title": "ApiResponse[CollectionVerifySingleData]"
      },
...
      "CollectionAddData": {
        "properties": {
          "created": {
            "type": "boolean",
            "title": "Created"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "created"
        ],
        "title": "CollectionAddData",
        "description": "添加收藏数据"
      },
      "CollectionAddRequest": {
        "properties": {
          "tmdb_id": {
            "type": "integer",
            "title": "Tmdb Id",
            "description": "TMDB ID"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "description": "媒体类型: movie 或 tv"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "标题"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year",
            "description": "年份"
          },
          "poster_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Path",
            "description": "海报路径"
          },
          "backdrop_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Path",
            "description": "背景图路径"
          },
          "share_url": {
            "type": "string",
            "title": "Share Url",
            "description": "夸克分享链接"
          },
          "share_pwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Pwd",
            "description": "分享密码"
          },
          "file_structure": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "File Structure",
            "description": "文件结构 (JSON)"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "分类: movie, tv, anime, documentary"
          }
        },
        "type": "object",
        "required": [
          "tmdb_id",
          "media_type",
          "title",
          "share_url"
        ],
        "title": "CollectionAddRequest",
        "description": "添加收藏的请求体",
        "example": {
          "media_type": "movie",
          "poster_path": "/poster.jpg",
          "share_url": "https://pan.quark.cn/s/xxxxx",
          "title": "盗梦空间",
          "tmdb_id": 27205,
          "year": 2010
        }
      },
      "CollectionCheckData": {
        "properties": {
          "collected": {
            "type": "boolean",
            "title": "Collected"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "collected"
        ],
        "title": "CollectionCheckData",
        "description": "检查收藏状态数据"
      },
      "CollectionCheckLinkData": {
        "properties": {
          "collected": {
            "type": "boolean",
            "title": "Collected"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "collected"
        ],
        "title": "CollectionCheckLinkData",
        "description": "检查链接收藏状态数据"
      },
      "CollectionCheckLinksData": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CollectionLinkStatus"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "CollectionCheckLinksData",
        "description": "批量检查链接收藏状态数据"
      },
      "CollectionCheckLinksRequest": {
        "properties": {
          "links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Links",
            "description": "分享链接列表"
          }
        },
        "type": "object",
        "required": [
          "links"
        ],
        "title": "CollectionCheckLinksRequest",
        "description": "批量检查链接收藏状态请求"
      },
      "CollectionDeleteData": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "CollectionDeleteData",
        "description": "删除收藏数据"
      },
      "CollectionItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tmdb_id": {
            "type": "integer",
            "title": "Tmdb Id"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "poster_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Path"
          },
          "backdrop_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Path"
          },
          "quark_share_url": {
            "type": "string",
            "title": "Quark Share Url"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "default": 0
          },
          "saved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Saved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tmdb_id",
          "media_type",
          "title",
          "quark_share_url",
          "saved_at"
        ],
        "title": "CollectionItem",
        "description": "收藏列表项"
      },
...
      "CollectionListData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "CollectionListData",
        "description": "收藏列表数据"
      },
      "CollectionVerifyRequest": {
        "properties": {
          "collection_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Ids",
            "description": "可选，仅验证这些收藏 ID（为空时验证全部状态为 1/3 的收藏）"
          }
        },
        "type": "object",
        "title": "CollectionVerifyRequest",
        "description": "批量验证收藏网盘状态请求"
      },
...
      "CollectionVerifySingleData": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/CollectionVerifyResult"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "CollectionVerifySingleData",
        "description": "单条收藏网盘状态验证数据"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "SseEventData": {
        "type": "object",
        "required": [
          "type",
          "current",
          "total",
          "percentage",
          "message",
          "level"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "current": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "percentage": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "level": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "SseEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SseEventData"
          }
        }
      }
    }
  }
}